+++ multiversx_sdk_core/__init__.py
@@ -4,8 +4,8 @@ from multiversx_sdk_core.address import (Address, AddressConverter,
 from multiversx_sdk_core.code_metadata import CodeMetadata
 from multiversx_sdk_core.contract_query import ContractQuery
 from multiversx_sdk_core.contract_query_builder import ContractQueryBuilder
-from multiversx_sdk_core.legacy_message import LegacyMessage
 from multiversx_sdk_core.message import Message
+from multiversx_sdk_core.message_v1 import MessageV1
 from multiversx_sdk_core.token_payment import TokenPayment
 from multiversx_sdk_core.transaction import Transaction
 from multiversx_sdk_core.transaction_payload import TransactionPayload
@@ -13,7 +13,7 @@ from multiversx_sdk_core.transaction_payload import TransactionPayload
 __all__ = [
     "AccountNonceHolder", "Address", "AddressConverter", "AddressFactory",
     "Transaction", "TransactionPayload",
-    "Message", "LegacyMessage",
+    "Message", "MessageV1",
     "CodeMetadata", "TokenPayment",
     "ContractQuery", "ContractQueryBuilder"
 ]
